+++ modelzoo/Recommendation/criteo_ctr/train.py
@@ -16,17 +16,16 @@
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
-import sys
+
 import tensorflow as tf
 from absl import app
 from absl import flags
-from tensorflow.keras import backend as K
 from tensorflow_recommenders_addons import dynamic_embedding as de
 
 from dcn_v2 import Ranking
 from deepray.core.base_trainer import Trainer
 from deepray.datasets.criteo import CriteoTsvReader
-from deepray.utils.export.export import export_to_savedmodel, export_for_serving
+from deepray.utils.export import export_to_savedmodel
 
 FLAGS = flags.FLAGS
 
@@ -45,21 +44,6 @@ def main(_):
   trainer.fit(train_input=train_input_fn, steps_per_epoch=FLAGS.steps_per_epoch)
 
   export_to_savedmodel(trainer.model)
-  if FLAGS.use_horovod:
-    FLAGS([
-        sys.argv[0],
-        "--use_horovod=False",
-    ])
-
-    K.clear_session()
-    de.enable_inference_mode()
-
-    model = Ranking(interaction="cross")
-    FLAGS([
-        sys.argv[0],
-        "--use_horovod=True",
-    ])
-    export_for_serving(trainer.model, export_model=model)
 
 
 if __name__ == "__main__":
